requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
mysql-connector-python==8.1.0
numpy==1.26.0
scikit-learn==1.3.1
scipy==1.15.3
python-dotenv==1.0.0
geopy==2.4.0
tensorflow==2.13.0  # 可选依赖 